+++ FishingBooker/Models/LoyaltyScaleViewModel.cs
@@ -20,17 +20,17 @@ namespace FishingBooker.Models
 
         [Display(Name = "Clients benefits for this scale(%)")]
         [Range(1, 100, ErrorMessage = "Percentage must be between 1 and 100!")]
-        public float ClientsBenefits { get; set; }
+        public int ClientsBenefits { get; set; }
 
 
         [Display(Name = "Owners benefits for this scale(%)")]
         [Range(1, 100, ErrorMessage = "Percentage must be between 1 and 100!")]
-        public float OwnerBenefits { get; set; }
+        public int OwnerBenefits { get; set; }
 
 
         [Display(Name = "Minumum earned points for a scale")]
         [Range(0, 999999, ErrorMessage = "Range is between 0 and 999999")]
         [Required(ErrorMessage = "You need to enter a number.")]
-        public float MinEarnedPoints { get; set; }
+        public int MinEarnedPoints { get; set; }
     }
 }
