+++ FishingBookerLibrary/FishingBookerLibrary/Models/LoyaltyScale.cs
@@ -11,8 +11,8 @@ namespace FishingBookerLibrary.Models
         public int Id { get; set; }
         public int LoyaltyProgramId { get; set; }
         public string ScaleName { get; set; }
-        public float ClientsBenefits { get; set; }
-        public float OwnerBenefits { get; set; }
-        public float MinEarnedPoints { get; set; }
+        public int ClientsBenefits { get; set; }
+        public int OwnerBenefits { get; set; }
+        public int MinEarnedPoints { get; set; }
     }
 }
